+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.1.6
 
-- Fix handle attribute error when supplier return internal error with no status code.
+- Fix handle attribute error when supplier return internal error with status code 200.
 
 ## 0.1.5